[refactor] 변수명 변경

+++ src/main/java/com/hanghae/reviewservice/dto/ProductRequestDto.java
@@ -5,6 +5,6 @@ import lombok.Getter;
 @Getter
 public class ProductRequestDto {
     private Long reviewCount;
-    private double score;
+    private float score;
 
 }
